clean up code to match python 3.6 style
To match python 3.6 style Alert has to inherit the ABC meta class
directly in the definition. But on the same time Alert is a subclass of
QObject which itself has an ABC meta class. Thus to avoid the "TypeError: metaclass conflict:" a new AlertMeta class had to be created.

See: http://www.phyast.pitt.edu/~micheles/python/metatype.html